extract storage logic to prep comparing storage options
