+++ src/Lorentz/Contracts/Permit/Type.hs
@@ -9,11 +9,23 @@
 module Lorentz.Contracts.Permit.Type where
 
 import Lorentz
-import Michelson.Text
 import Tezos.Crypto.Hash
 import Util.Named
+import Michelson.Parser -- hiding (parseValue)
+import Michelson.TypeCheck.Instr
+import Michelson.TypeCheck.TypeCheck
+import Michelson.Macro
 
+import Data.Either
+import Data.Function
+import Control.Applicative
+import Control.Monad hiding ((>>), fail)
+import Control.Monad.Fail
 import Text.Show (Show(..))
+
+import Control.Monad.Trans.Reader
+import Data.Singletons
+import Text.Megaparsec (eof)
 
 newtype Blake2B =
   UnsafeBlake2B ByteString
@@ -61,36 +73,6 @@ instance HasTypeAnn Permit
 toPermit :: (Blake2B, Address) & s :-> Permit & s
 toPermit = forcedCoerce_
 
-type Permits = BigMap Permit ()
-
--- | Add a `Permit` to `Permits`
-addToPermits :: Permit & Permits & s :-> Permits & s
-addToPermits = do
-  dip $ do
-    unit
-    some
-  update
-
--- | Assert the parameter exists and deletes it from `Permits`
-assertSentParam :: forall s. Permits & Permit & s :-> Permits & s
-assertSentParam = do
-  swap
-  pair
-  dup
-  unpair
-  mem @Permits
-  if Holds
-     then do
-       dup
-       car
-       dip $ do
-         cdr
-         none
-       update
-     else do
-       push $ mkMTextUnsafe "no permit"
-       failWith
-
 data Parameter cp
   = Permit !SignedParams
   | Wrapped !cp
@@ -102,23 +84,34 @@ deriving anyclass instance (IsoValue cp) => IsoValue (Parameter cp)
 instance (HasTypeAnn cp, IsoValue cp) => ParameterHasEntryPoints (Parameter cp) where
   type ParameterEntryPointsDerivation (Parameter cp) = EpdPlain
 
-data Storage st = Storage
-  { presignedParams :: !Permits
+data Storage permits st = Storage
+  { presignedParams :: !permits
   , counter         :: !("counter" :! Natural)
   , wrappedStorage  :: !st
   }
   deriving stock Generic
 
-deriving stock instance (Show st) => Show (Storage st)
-deriving anyclass instance (IsoValue st) => IsoValue (Storage st)
+deriving stock instance (Show permit, Show st) => Show (Storage permit st)
+deriving anyclass instance (IsoValue permit, IsoValue st) => IsoValue (Storage permit st)
 
-unStorage :: Storage st & s :-> (Permits, ("counter" :! Natural, st)) & s
+unStorage :: Storage permits st & s :-> (permits, ("counter" :! Natural, st)) & s
 unStorage = forcedCoerce_
 
-toStorage :: (Permits, ("counter" :! Natural, st)) & s :-> Storage st & s
+toStorage :: (permits, ("counter" :! Natural, st)) & s :-> Storage permits st & s
 toStorage = forcedCoerce_
 
-mkStorage :: st -> Storage st
+incrementCounter :: ("counter" :! Natural, st) & s :-> ("counter" :! Natural, st) & s
+incrementCounter = do
+  dup
+  car
+  forcedCoerce_ @("counter" :! Natural) @Natural
+  push @Natural 1
+  add
+  forcedCoerce_ @Natural @("counter" :! Natural)
+  dip cdr
+  pair
+
+mkStorage :: Monoid permits => st -> Storage permits st
 mkStorage = Storage mempty (#counter .! 0)
 
 -- | `Storage` we can parse from the Tezos RPC
@@ -132,4 +125,26 @@ data DummyStorage st = DummyStorage
 deriving stock instance (Show st) => Show (DummyStorage st)
 deriving anyclass instance (IsoValue st) => IsoValue (DummyStorage st)
 
+fromDummyStorage :: forall permits st. Monoid permits => DummyStorage st -> Storage permits st
+fromDummyStorage DummyStorage{..} =
+  Storage mempty counter wrappedStorage
+
+-- | Parse and typecheck a Michelson value
+parseTypeCheckValue ::
+     forall t. (SingI t)
+  => Parser (Value t)
+parseTypeCheckValue =
+  (>>= either (fail . show) return) $
+  runTypeCheckIsolated . flip runReaderT def . typeCheckValue . expandValue <$>
+  (value <* eof)
+
+withAdmin42Storage :: forall permit t r. (Monoid permit, IsoValue t, SingI t) => Text -> (Storage permit t -> r) -> r
+withAdmin42Storage storageTxt f =
+  f . fromDummyStorage $ fromVal @(DummyStorage t) param
+  where
+    parsedParam = parseNoEnv
+      (parseTypeCheckValue @(ToT (DummyStorage t)))
+      "Permit DummyStorage"
+      storageTxt
+    param = either (error . fromString . show) id parsedParam
